🧬 Move the tests about utility function about handling JavaScript attributes into another test module.
